Enabled software SPI by copying extras/sw_spi/src/GxEPD2_EPD.{h,cpp} into src/
